remove unnecessary store folder and move store file to user folder
